Rename file to app-metadata.ts and small fixes
