+++ test/unit/project-management/ios-app.spec.ts
@@ -25,7 +25,7 @@ import { ProjectManagementRequestHandler } from '../../../src/project-management
 import { deepCopy } from '../../../src/utils/deep-copy';
 import { FirebaseProjectManagementError } from '../../../src/utils/error';
 import * as mocks from '../../resources/mocks';
-import { IosAppMetadata, AppPlatform } from '../../../src/project-management/AppMetadata';
+import { IosAppMetadata, AppPlatform } from '../../../src/project-management/app-metadata';
 
 const expect = chai.expect;
